Replaced the aligner and FIFO in the prefetchbuffer with the a module called cv32e40x_alignment_buffer.
This alignment buffer is a slighly modified version of the
old fetch_fifo as it was before the aligner+fifo was introduced.

FIFO_DEPTH must be set to 3 to not get additional stalls as compared
to the version before this change. Future optimizations to decrease
this to 1.5 words should be possible.

NOT SEC equivalent.

Signed-off-by: Oystein Knauserud <Oystein.Knauserud@silabs.com>