Remember XIF accept/reject in case ID is stalled
The ID stage must remember whether the coprocessor has already
accepted or rejected the instruction that is currently in the ID stage
if the EX stage is not ready (i.e., ID is stalled internally by
ex_ready_i being low).

An alternative would be to keep xif_issue_if.x_issue_valid low while
ex_ready_i is not asserted, but that would unnecessarily delay
offloading of the instruction.

Signed-off-by: Michael Platzer <michael.platzer@hotmail.com>